1.0.3
Part one of my uber refactoring, as in I'm totally over hauling, connectED.  I'll be integrating support for various 'social' networking system.  And well I'm doing a ton of work.  Right now this commit is a total of two days of work, especially my work toward implementing GtkBuilder an GObjects.  But for today I've done a ton & now: I must sleep.

:-) How cool is it that my web app revamp gives me an excuse to work on another of my C projects. . . One I've let sit for 'to long'.  Which makes it more fun to work on and yet kinda more gross.  I'd so forgotten how icky and just plain nasty most of this code is.  *sigh* such is the beauty of OSS.  Forking means 'potential' doesn't have to die...but bad code & bed design can.  Which is kinda fun & kinda mind-numbing.

Oh this is so gonna be a blog post... LOL of course its already 3 podcast segments so I may just leave it at that.

But that's for tomorrow and each tomorrow afterward.  Now to prepare for those tomorrows... well the 'next' one at least ;-)

Committed on : 2010-02-18 @ 08:56:50 PM

+++ src/classbrowser_parse.c
@@ -33,7 +33,7 @@ static gchar *read_text_file( gchar *filename )
 	FILE *fp;
 	gchar *buffer = NULL;
 	struct stat buf;
-	size_t size_read;
+	gint size_read;
 
 	stat(filename, &buf);
